Models: Support _set_*, _get_*, _filter_* methods.
Allows implementing setters, getters, and simple value modifiers.

For example, a _filter_date() method would ensure that the internal field value is a Date object, even when initialized as a string from JSON.

I don't even know how else to work with Date fields, so the lack of such a facility is quite irritating.

This is copied directly from the getflow.com source, and mentioned here:

http://www.quora.com/Flow-web-application/What-JavaScript-framework-does-Flow-use-for-its-web-frontend/answer/Jason-Webster-2